+++ extensions/pipe-commands/src/utils.ts
@@ -33,10 +33,10 @@ export async function parseScriptCommands(): Promise<{
   commands: ScriptCommand[];
   invalid: InvalidCommand[];
 }> {
-  let { pipeCommandsFolder = environment.supportPath } = getPreferenceValues<{ pipeCommandsFolder: string }>();
-  pipeCommandsFolder = untildify(pipeCommandsFolder);
+  const { commandFolder } = getPreferenceValues<{ commandFolder: string }>();
+
   const defaultPaths = globbySync(`${environment.assetsPath}/commands/**/*`);
-  const userPaths = globbySync(`${pipeCommandsFolder}/**/*`);
+  const userPaths = commandFolder ? globbySync(`${untildify(commandFolder)}/**/*`) : [];
   const scriptPaths = [...userPaths, ...defaultPaths].filter(
     (path) => !(path.startsWith(".") || path.endsWith(".png") || path.endsWith(".svg"))
   );
@@ -58,7 +58,7 @@ export async function parseScriptCommands(): Promise<{
   for (const command of commands) {
     const res = validator.validate(command.metadatas, schema);
     if (res.valid) {
-      valids.push({ ...command, user: command.path.startsWith(pipeCommandsFolder) });
+      valids.push({ ...command, user: command.path.startsWith(commandFolder) });
     } else {
       invalid.push({ path: command.path, content: command.content, errors: res.errors.map((err) => err.stack) });
     }